ui: remove titlemidlarge style
Remove the TitleMidLarge style, replacing it with more Material3
attributes.

This finally removes all of the non-standard text styles. There is
still the two TitleMedium styles, but this is considered okay.

In the playback view, TitleMidLarge has become TitleLarge, which
honestly works quite well, as the same text styles are used in the
detail views.